# great-circle distance speedups