--- a/siteupdate/cplusplus/classes/HighwaySegment/HighwaySegment.h
+++ b/siteupdate/cplusplus/classes/HighwaySegment/HighwaySegment.h
@@ -6,6 +6,7 @@ class HighwaySegment
 	Waypoint *waypoint1;
 	Waypoint *waypoint2;
 	Route *route;
+	double length;
 	std::list<HighwaySegment*> *concurrent;
 	std::unordered_set<TravelerList*> clinched_by;
 	std::mutex clin_mtx;
@@ -15,7 +16,6 @@ class HighwaySegment
 	std::string str();
 	bool add_clinched_by(TravelerList *);
 	std::string csv_line(unsigned int);
-	double length();
 	std::string segment_name();
 	unsigned int index();
 	//std::string concurrent_travelers_sanity_check();