### C++:  store & retrieve segment length
Increases speed slightly when Computing stats & Creating per-traveler stats logs.
Along for the ride: added alternate great-circle distance formulas to Waypoint::distance_to & PlaceRadius::contains_vertex, commented out for now