@yakra 2019-06-03
Added a commit to implement this in siteupdate.py.

### On noreaster:
* *Reading waypoints for all routes* 116.9 -> 120.1 s, gaining 3.2 s.
* *Computing stats* 30.1 -> 26.1 s, losing 4.0 s.
Let's call this moving 3.6 s +/- 0.4 s from point A to point B. Times always seem to vary a little bit between runs in Python *(for example, the total run time for all tasks other than the ones I measured for this benchmark went 260.2 -> 254.8 s)*; there will be a little noise in the signal.
* *Creating per-traveler stats logs and augmenting data structure* 46.2 -> 29.9 s. A 50% speedup for this task.