@yakra 2019-06-03
### Spherical law of cosines formula for area graphs
253f75270dfee3650d05a5962e44074e7284f5a2 http://forum.travelmapping.net/index.php?topic=3014.msg13996#msg13996

Same as the existing formula, but with some terms factored out or simplified via trig identity.
Increases area graph generation speed by ~32%.
* Not such a big deal in C++ on BiggaTomato, where it gets us from 3.7 to 2.8 s.
* In Python on noreaster, we go from 56.8 to 43.1.

No DIFFs in the resulting graphs.